Handle multiple JSON Vulnerability Prefixes in same response
An error ("Unable to parse JSON string") occurs when the same batch
response contains multiple JSON vulnerability prefixes (such as when
multiple GET operations that return arrays are batched).  A unit test
was added to illustrate this scenario.

This occurred only with multiple prefixes because string.replace was
used on the entire response text, and string.replace only replaces the
first occurrence of the substring.

To fix this, I moved the trim function to the "httpAdapter", because
that allowed finer control over when the prefix was trimmed - the prefix
is now trimmed for each batch response separately and only when parsing
a response with a 'json' content type.  Further, trimming JSON
vulnerability prefixes don't appear relevant for multifetch:  the
returned data is in an object form, not an array - so the vulnerability
does not apply (see this for more details of how the vulnerability
manifests:
http://haacked.com/archive/2008/11/20/anatomy-of-a-subtle-json-vulnerability.aspx/
).